Main Server 필요없는 함수 제거 및 병합

+++ The Eternals/FrameServer/Server/Header/NPCAnimation.h
@@ -7,7 +7,7 @@ namespace NPC_TYPE
 	constexpr int IDLE		= 1;
 	constexpr int WALK		= 2;
 
-	constexpr double TPS	= 4800.0;
+	constexpr float TPS	= 4800.f;
 }
 
 namespace NPC_NORMAL_TYPE
